chore: move scheduler of batchupdate to websiteupdate
