--- backend/internal/jobs/batchupdate/params.go
@@ -1,3 +1,0 @@
-package batchupdate
-
-type Params struct{}
